+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 <font color=red><strong>删除</strong></font>
 
 
+# 3.0.0
+1. <font color=green>新增</font> `globalVarFile` 配置 `less` 变量文件路径，文件中的 `less` 变量会从所有less文件中把对应css变量替换成less变量
+2. <font color=green>新增</font> `findMissVar` 配置 `globalVarFile` 文件中与主题色相关的变量引用了其它文件的变量且该变量不是 `varFile` 中定义的变量时，是否要把引用的变量加入到主题色变量里来，防止编译时报错。
+
+
 # 2.0.0
 1. <font color=blue>增强</font> 支持webpack 5